requirements.txt
asgiref==3.11.0
attrs==25.4.0
diff-match-patch==20241021
Django==5.2.10
django-import-export==4.3.14
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
pillow==12.0.0
PyJWT==2.10.1
python-dotenv==1.2.1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
tablib==3.9.0
tzdata==2025.3
uritemplate==4.2.0
psycopg[binary]>=3.1